requirements.txt
-e .
aws_cdk.aws_s3==1.92.0

black==20.8b1
pre-commit==2.9.3
boto3
botocore
psycopg2-binary
